Moved the fleets.js file to the models directory
- Moved the `fleet.js` file to the models directory
- Updated the reference for fleets in `apiController.js`, `commanderController.js`, `fleetManagementController.js`, `pagesController` and `index.js`
