+++ controllers/apiController.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const setup = require('../setup.js');
 const cache = require('../cache.js')(setup);
-const fleets = require('../fleets.js')(setup);
+const fleets = require('../models/fleets.js')(setup);
 const user = require('../user.js')(setup);
 const users = require('../users.js')(setup);
 const refresh = require('passport-oauth2-refresh');
