+++ index.js
@@ -36,7 +36,7 @@ database.connect(function () {
 
 	const users = require('./users.js')(setup);
 	const customSSO = require('./customSSO.js')(refresh, setup, request, url);
-	const fleets = require('./fleets.js')(setup);
+	const fleets = require('./models/fleets.js')(setup);
 	const waitlist = require('./globalWaitlist.js')(setup);
 
 	//Start timers
